requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Data fetching
databento>=0.30.0

# Machine Learning and Analysis
qlib>=0.9.0
lightgbm>=3.3.0
xgboost>=1.6.0

# Backtesting
backtrader>=1.9.76

# Visualization
plotly>=5.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Web Interface
gradio>=4.0.0
fastapi>=0.115.2,<1.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.20
starlette>=0.40.0,<1.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=0.19.0

# Data storage
pyarrow>=8.0.0
h5py>=3.7.0

# Development and testing
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

# Optional dependencies for enhanced functionality
# jupyter>=1.0.0
# ipywidgets>=7.6.0
# ta-lib>=0.4.24  # Technical analysis library
# yfinance>=0.1.70  # Alternative data source
# ccxt>=2.0.0  # Cryptocurrency exchange data
